+++ app_page/saysafe_app_page/phone/i_declare_peace_page.py
@@ -1,0 +1,16 @@
+# coding=utf-8
+#我要报平安页面
+from public.selenium_page import SeleniumPage
+
+
+class IDeclarePeacePage(SeleniumPage):
+
+    # 页面元素
+
+    # 添加按钮
+
+    addButtonLoc = ".add_btn"
+
+
+
+
