+++ qiqiao_page/mobile_page/mobile_application_page.py
@@ -1,0 +1,1 @@
+# coding=utf-8
