+++ qiqiao_page/mobile_page/mobile_home_page.py
@@ -1,0 +1,1 @@
+# coding=utf-8
