+++ readme.md
@@ -1,0 +1,7 @@
+命名规范：
+        包命名：不能含有大写字母：已单个单词加“_”下划线的形式。如file_data
+        类命名: 大驼峰写法，单个单词首字母大写。如SeleniumPage
+        函数命名：大驼峰写法，单个单词首字母大写。如GetUserName()
+        变量命名：小驼峰法，首个单词首字母小写。如myStudentCount
+
+ 目录结构说明：
