+++ test.py
@@ -1,9 +1,8 @@
 import time
-from public.driver import pcdriver
-from qiqiao_PublicPage.login_page import LoginPage
-from qiqiao_PublicPage.portal_page import PortalPage
-from qiqiao_PublicPage.applicationList_page import ApplicationListPage
-from AppPage_obj.SaySafePage.fill_data_page import AppFillSafe
+from qiqiao_page.pc_page.login_page import LoginPage
+from qiqiao_page.pc_page.portal_page import PortalPage
+from qiqiao_page.pc_page.applicationList_page import ApplicationListPage
+from app_page.saysafe_app_page.pc.fill_data_page import AppFillSafe
 from selenium import webdriver
 
 driver = webdriver.Chrome(executable_path="D:\Projects\qiqiao_autoTest\\file_data\chromedriver.exe")
