+++ testcase/saysafe_app_pc_testcase/filldata.py
@@ -1,10 +1,10 @@
 import time
 import unittest
 from public.driver import pcdriver
-from qiqiao_PublicPage.login_page import LoginPage
-from qiqiao_PublicPage.portal_page import PortalPage
-from qiqiao_PublicPage.applicationList_page import ApplicationListPage
-from AppPage_obj.SaySafePage.fill_data_page import AppFillSafe
+from qiqiao_page.pc_page.login_page import LoginPage
+from qiqiao_page.pc_page.portal_page import PortalPage
+from qiqiao_page.pc_page.applicationList_page import ApplicationListPage
+from app_page.saysafe_app_page.pc.fill_data_page import AppFillSafe
 
 
 class FillReportTest(unittest.TestCase):
